+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 TODO: Date
 
 * **Breaking:**
-  * Increased minimum supported Rust version from 1.45.0 to 1.46.0.
+  * Increased minimum supported Rust version from 1.45.0 to 1.50.
   * Removed "rhizome" features (always enabled now)
   * Removed "styles" and "topiary" features. CSS scoping will be enabled through more general means.
   * Reworked generated component interface
@@ -39,6 +39,15 @@ TODO: Date
     ```
 
   * Optional arguments: `pattern?: Type`
+  * Repeat arguments: `item_name/pattern*: Type` or `item_name/pattern+: Type`
+
+    The `item_name/` is optional.
+
+    Repeat arguments work in combination with `?` and/or defaults.
+
+    When using `*?`, you can flatten the resulting `Option<Vec<_>>` into just `Vec<_>` by writing `*?.flatten`.  
+    The default value defaults to `Vec::default()` in this case.
+
   * Default parameters: `pattern: Type = default`
   * Conditional attributes: `."attribute-name"? = {Option<&'bump str>}`
   * Conditional parameters (like conditional attributes)